athena: fix tables and remove unnecessary promises
* Fixed tables() so that it doesn't return the column name table_names.

* Avoided the creation of unnecessary promises.

* Used slice() and map() to generate the response rows.
